Christian Walther      2008-01-13 07:59:42 PST
I'm trying to build a set of static SDL_* libraries in my Mingw
cross-compilation setup on Mac OS X. I have only static (no dynamic) versions
of libpng and libtiff installed and configure SDL_image using

cross-configure.sh --prefix=/usr/local/cross-tools/i386-mingw32
--disable-shared --disable-jpg-shared --disable-png-shared --disable-tif-shared

In this invocation, the checks for libpng and libtiff fail (even though the
libraries are there, and are found) because of missing symbols from zlib. I can
fix this using the attached patch. I assume this shouldn't interfere in the
case that dynamic png/tiff libraries are around, but I'm reluctant to try it,
and I'm not very familiar with autoconf in general. Perhaps someone who is can
advise if this is the right thing to do...