SDL_image uses ImageIO for ios4 and osx and UIImage on ios3
Vittorio Giovara

>> The second one is a feature request to make it so that when the
>> ImageIO framework is not present the library would fallback to the
>> UIImage backend; although it would perhaps make backporting a little
>> difficult I am sure that there is still a lot of software targeting
>> pre-ios 4 that could make use of it.
>
> It's possible, and not entirely unreasonable, but iOS moves so fast
> that I haven't felt it really merited the extra effort to do so. My
> experience is that there isn't a lot of interest in 3.x. I have a
> 3.1.3 device and there are actually very few apps (that I would be
> interested in getting) that support 3.x any more and I have to be very
> careful to no upgrade my existing apps because the new versions don't
> work on my device any more. I think the percentage of devices 3.x is
> less than 10% now. Testing old devices is another difficulty. And if
> you need to start weak linking frameworks in the Xcode project, you go
> into areas most non Apple developers don't know about.
>
> -Eric

I see your point, iOS is a much faster platform than OSX and a lot of
applications are dropping support for 3.* devices, however I feel that
as SDL is a compatibility layer between architectures it should be as
compatible as possible, until major api change or build system make it
impossible/unbearable to keep it updated.
Having said that I am proposing a solution that should keep SDL_image
compatible with both the newer ImageIO for ios4 and osx and allow us
to use uiimage on ios3: what I am linking is a merge of
IMG_UIImage.m and IMG_ImageIO.c and the small changes that need to be
applied to the related xcode project files).